demos/xrayedge/containers/image-generator/requirements.txt
boto3==1.9.*
mysql-connector-python==8.0.*
